Feat Store :
- Réorganisation de Activity_store : suppresion des slides et des fragments qui vont avec.
- Ajout du bouton flottant pour le QRCode
