--- bee/src/main/java/com/apisense/bee/ui/fragment/CategoryStoreFragment.java
@@ -1,22 +1,0 @@
-package com.apisense.bee.ui.fragment;
-
-import android.os.Bundle;
-import android.support.v4.app.Fragment;
-import android.view.LayoutInflater;
-import android.view.View;
-import android.view.ViewGroup;
-
-import com.apisense.bee.R;
-
-
-public class CategoryStoreFragment extends Fragment {
-
-    @Override
-    public View onCreateView(LayoutInflater inflater, ViewGroup container,
-                             Bundle savedInstanceState) {
-
-        View root = inflater.inflate(R.layout.fragment_store_category, container, false);
-
-        return root;
-    }
-}
