--- bee/src/main/res/layout/fragment_store_category.xml
@@ -1,80 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<LinearLayout xmlns:android="http://schemas.android.com/apk/res/android"
-    android:layout_width="match_parent"
-    android:layout_height="match_parent"
-    android:orientation="vertical">
-
-
-    <RelativeLayout xmlns:android="http://schemas.android.com/apk/res/android"
-        android:layout_width="match_parent"
-        android:layout_height="wrap_content"
-        android:orientation="horizontal">
-
-        <Spinner
-            android:id="@+id/spinner"
-            android:layout_width="wrap_content"
-            android:layout_height="wrap_content" />
-
-        <RelativeLayout
-            android:id="@+id/item"
-            android:layout_width="match_parent"
-            android:layout_height="wrap_content"
-            android:orientation="vertical">
-
-            <LinearLayout
-                android:id="@+id/thumbnail"
-                android:layout_width="wrap_content"
-                android:layout_height="wrap_content"
-                android:layout_alignParentLeft="true"
-                android:layout_alignParentStart="true"
-                android:layout_gravity="center_vertical"
-                android:layout_margin="@dimen/thumbnail_margin">
-
-                <ImageView
-                    android:id="@+id/home_item_icon"
-                    android:layout_width="60dp"
-                    android:layout_height="60dp"
-                    android:background="@drawable/ic_launcher_bee"
-                    android:contentDescription="@string/experiment_image_desc" />
-
-            </LinearLayout>
-
-
-            <LinearLayout
-                android:id="@+id/experiment_header"
-                android:layout_width="match_parent"
-                android:layout_height="wrap_content"
-                android:layout_alignBottom="@+id/thumbnail"
-                android:layout_toEndOf="@id/thumbnail"
-                android:layout_toRightOf="@id/thumbnail"
-                android:orientation="vertical"
-                android:weightSum="1">
-
-
-                <View
-                    android:layout_width="fill_parent"
-                    android:layout_height="1dp"
-                    android:layout_marginLeft="10dp"
-                    android:layout_marginStart="10dp"
-                    android:layout_marginTop="32dp"
-
-                    android:background="@color/light_grey" />
-
-            </LinearLayout>
-
-            <com.apisense.bee.widget.ApisenseTextView
-                style="@style/ApisenseSlideTitleText"
-                android:layout_width="wrap_content"
-                android:layout_height="wrap_content"
-                android:layout_alignLeft="@+id/experiment_header"
-                android:layout_alignStart="@+id/experiment_header"
-                android:layout_centerVertical="true"
-                android:layout_weight="1.03"
-                android:text="@string/store_section_general" />
-
-        </RelativeLayout>
-
-
-    </RelativeLayout>
-
-</LinearLayout>
