--- bee/src/main/res/layout/fragment_store_home.xml
@@ -1,35 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<LinearLayout xmlns:android="http://schemas.android.com/apk/res/android"
-    android:layout_width="match_parent"
-    android:layout_height="match_parent"
-    android:orientation="vertical">
-    <!-- Experiments list -->
-    <ListView
-        android:id="@+id/store_experiment_lists"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content" />
-
-    <!-- The menu elements ListView. -->
-    <ListView
-        android:id="@+id/left_drawer"
-        android:layout_width="match_parent"
-        android:layout_height="match_parent"
-        android:layout_gravity="start"
-        android:background="@drawable/grandient_background"
-        android:choiceMode="singleChoice"
-        android:divider="@android:color/transparent"
-        android:dividerHeight="0dp" />
-
-    <LinearLayout
-        android:id="@+id/store_empty_list"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:layout_margin="@dimen/activity_horizontal_margin">
-
-        <com.apisense.bee.widget.ApisenseTextView
-            android:layout_width="wrap_content"
-            android:layout_height="wrap_content"
-            android:gravity="center"
-            android:text="@string/no_available_experiments" />
-    </LinearLayout>
-</LinearLayout>
